requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
docopt==0.6.2
idna==3.10
iniconfig==2.0.0
packaging==24.2
pillow==11.1.0
pipreqs==0.4.13
pipreqsnb==0.2.4
pluggy==1.5.0
pytest==8.3.5
requests==2.32.3
setuptools==75.6.0
svgwrite==1.4.3
urllib3==2.3.0
yarg==0.1.10
